feat:cli and modules done
